Consolidate services specs. Uninstall Console if removed from definition
This PR adds validation to the services we deploy in case the metadata for services or exposure changes on the tenant spec, this will update the services.

This also uninstalls console if it's removed.

Lastly, I refactored a lot of the code used for these checks into their own functions on separate files as main-controller.go number of lines is getting out of control
