+++ pkg/controller/cluster/main-controller.go
@@ -773,8 +773,6 @@ func (c *Controller) syncHandler(key string) error {
 	uOpts := metav1.UpdateOptions{}
 	gOpts := metav1.GetOptions{}
 
-	var consoleDeployment *appsv1.Deployment
-
 	// Convert the namespace/name string into a distinct namespace and name
 	if key == "" {
 		runtime.HandleError(fmt.Errorf("Invalid resource key: %s", key))
@@ -850,23 +848,11 @@ func (c *Controller) syncHandler(key string) error {
 		}
 	}
 
-	// Handle the Internal ClusterIP Service for Tenant
-	svc, err := c.serviceLister.Services(tenant.Namespace).Get(tenant.MinIOCIServiceName())
+	// validate the minio service
+	err = c.checkMinIOSvc(ctx, tenant, nsName)
 	if err != nil {
-		if k8serrors.IsNotFound(err) {
-			if tenant, err = c.updateTenantStatus(ctx, tenant, StatusProvisioningCIService, 0); err != nil {
-				return err
-			}
-			klog.V(2).Infof("Creating a new Cluster IP Service for cluster %q", nsName)
-			// Create the clusterIP service for the Tenant
-			svc = services.NewClusterIPForMinIO(tenant)
-			_, err = c.kubeClientSet.CoreV1().Services(tenant.Namespace).Create(ctx, svc, cOpts)
-			if err != nil {
-				return err
-			}
-		} else {
-			return err
-		}
+		klog.V(2).Infof("Error when consolidating tenant service: %v", err)
+		return err
 	}
 
 	// Handle the Internal Headless Service for Tenant StatefulSet
@@ -1253,104 +1239,11 @@ func (c *Controller) syncHandler(key string) error {
 
 	}
 
-	if tenant.HasConsoleEnabled() {
-		// Get the Deployment with the name specified in MirrorInstace.spec
-		if consoleDeployment, err = c.deploymentLister.Deployments(tenant.Namespace).Get(tenant.ConsoleDeploymentName()); err != nil {
-			if !k8serrors.IsNotFound(err) {
-				return err
-			}
-			var userCredentials []*v1.Secret
-			if tenant.Spec.Users != nil {
-				for _, credential := range tenant.Spec.Users {
-					credentialSecret, sErr := c.kubeClientSet.CoreV1().Secrets(tenant.Namespace).Get(ctx, credential.Name, gOpts)
-					if sErr == nil && credentialSecret != nil {
-						userCredentials = append(userCredentials, credentialSecret)
-					}
-				}
-			}
-			if tenant.HasCredsSecret() && tenant.HasConsoleSecret() {
-				consoleSecretName := tenant.Spec.Console.ConsoleSecret.Name
-				consoleSecret, sErr := c.kubeClientSet.CoreV1().Secrets(tenant.Namespace).Get(ctx, consoleSecretName, gOpts)
-				if sErr == nil && consoleSecret != nil {
-					_, accessKeyExist := consoleSecret.Data["CONSOLE_ACCESS_KEY"]
-					_, secretKeyExist := consoleSecret.Data["CONSOLE_SECRET_KEY"]
-					if accessKeyExist && secretKeyExist {
-						userCredentials = append(userCredentials, consoleSecret)
-					}
-				} else {
-					// just log the error
-					klog.Info(sErr)
-				}
-			}
-			if len(userCredentials) == 0 {
-				msg := "Please set the credentials"
-				klog.V(2).Infof(msg)
-				if _, terr := c.updateTenantStatus(ctx, tenant, msg, totalReplicas); terr != nil {
-					return err
-				}
-				// return nil so we don't re-queue this work item
-				return nil
-			}
-			// Make sure that MinIO is up and running to enable MinIO console user.
-			if !tenant.MinIOHealthCheck() {
-				if _, err = c.updateTenantStatus(ctx, tenant, StatusWaitingForReadyState, totalReplicas); err != nil {
-					return err
-				}
-				return ErrMinIONotReady
-			}
-
-			if tenant, err = c.updateTenantStatus(ctx, tenant, StatusProvisioningConsoleDeployment, totalReplicas); err != nil {
-				return err
-			}
-
-			skipCreateConsoleUser := false
-			// If Console is deployed with the CONSOLE_LDAP_ENABLED="on" configuration that means MinIO is running with LDAP enabled
-			// and we need to skip the console user creation
-			for _, env := range tenant.GetConsoleEnvVars() {
-				if env.Name == "CONSOLE_LDAP_ENABLED" && env.Value == "on" {
-					skipCreateConsoleUser = true
-					break
-				}
-			}
-
-			if pErr := tenant.CreateConsoleUser(adminClnt, userCredentials, skipCreateConsoleUser); pErr != nil {
-				klog.V(2).Infof(pErr.Error())
-				return pErr
-			}
-
-			// Create Console Deployment
-			consoleDeployment = deployments.NewConsole(tenant)
-			_, err = c.kubeClientSet.AppsV1().Deployments(tenant.Namespace).Create(ctx, consoleDeployment, cOpts)
-			if err != nil {
-				klog.V(2).Infof(err.Error())
-				return err
-			}
-			// Create Console service
-			consoleSvc := services.NewClusterIPForConsole(tenant)
-			_, err = c.kubeClientSet.CoreV1().Services(svc.Namespace).Create(ctx, consoleSvc, cOpts)
-			if err != nil {
-				klog.V(2).Infof(err.Error())
-				return err
-			}
-		} else {
-
-			// Verify if this console deployment matches the spec on the tenant (resources, affinity, sidecars, etc)
-			consoleDeploymentMatchesSpec, err := consoleDeploymentMatchesSpec(tenant, consoleDeployment)
-			if err != nil {
-				return err
-			}
-
-			// if the console deployment doesn't match the spec
-			if !consoleDeploymentMatchesSpec {
-				if tenant, err = c.updateTenantStatus(ctx, tenant, StatusUpdatingConsole, totalReplicas); err != nil {
-					return err
-				}
-				consoleDeployment = deployments.NewConsole(tenant)
-				if _, err = c.kubeClientSet.AppsV1().Deployments(tenant.Namespace).Update(ctx, consoleDeployment, uOpts); err != nil {
-					return err
-				}
-			}
-		}
+	// Check whether console is enabled or if it should be removed and the state of it's service
+	err = c.checkConsoleStatus(ctx, tenant, totalReplicas, adminClnt, cOpts, uOpts, nsName)
+	if err != nil {
+		klog.V(2).Infof("Error checking console state %v", err)
+		return err
 	}
 
 	if tenant.HasKESEnabled() && tenant.TLS() {
@@ -1639,24 +1532,6 @@ func (c *Controller) handleObject(obj interface{}) {
 		c.enqueueTenant(tenant)
 		return
 	}
-}
-
-func (c *Controller) checkAndCreateConsoleCSR(ctx context.Context, nsName types.NamespacedName, tenant *miniov2.Tenant) error {
-	if _, err := c.certClient.CertificateSigningRequests().Get(ctx, tenant.ConsoleCSRName(), metav1.GetOptions{}); err != nil {
-		if k8serrors.IsNotFound(err) {
-			if tenant, err = c.updateTenantStatus(ctx, tenant, StatusWaitingConsoleCert, 0); err != nil {
-				return err
-			}
-			klog.V(2).Infof("Creating a new Certificate Signing Request for Console Server Certs, cluster %q", nsName)
-			if err = c.createConsoleTLSCSR(ctx, tenant); err != nil {
-				return err
-			}
-			// we want to re-queue this tenant so we can re-check for the console certificate
-			return errors.New("waiting for console cert")
-		}
-		return err
-	}
-	return nil
 }
 
 // MinIOControllerRateLimiter is a no-arg constructor for a default rate limiter for a workqueue for our controller.
